WIP: we need a way to translate to line to byteposition
Either by preparsing for new lines and creating a lookup table or by
storing all newlines with position in Cursor and lookup it up via
Tokenizer.

I would rather not extend Token as this is a very niche scenario which
is not worth passing the byte positions around.

Currently I prefer to create a specialized iterator that keeps track of
previous location and enhances it with a range by tracking the the line
numbers from the beginning.